Moved views to the separate files.
