+++ Trippy/Packages/TimelineView/Sources/TimelineView/TimelinePoint.swift
@@ -6,7 +6,6 @@
 //
 
 import UIKit
-import TimelineView
 import TrippyUI
 import Stevia
